Responsive comment load if smaller than breakpoint
Now you can specify to load comments in if smaller than, as well as larger than breakpoints. 

This is done with the data attribute names in triggerMin & triggerMax. 

This would load if larger than the specified breakpoint:
<div data-usn-load-if-larger="usn-sml-med">
    <!-- content -->
</div>

This would load if smaller than the specified breakpoint:
<div data-usn-load-if-smaller="usn-sml-med">
    <!-- content -->
</div>

I know the original Responsive Comments JS was created to be 'progressive enhancement only', but often there is a reason to load in different content & media on smaller devices.